Formatted email, added instructions how to manage device
